linux: Remove 3.5 (overo) and 3.6 (DuoVero) kernels
Overo and DuoVero are both supported by the mainline 3.15 kernel.
Binaries for hardware acceleration on these platforms are typically
not compatible with either of these versions.  As such we get rid of
these two kernels in favour of a recent mainline kernel and any
specific kernels required for binary compatibility.

Signed-off-by: Ash Charles <ashcharles@gmail.com>